+++ src/arize/_generated/api_client/test/test_ai_integration.py
@@ -49,7 +49,7 @@ class TestAiIntegration(unittest.TestCase):
                 enable_default_models = True,
                 function_calling_enabled = True,
                 auth_type = 'default',
-                provider_metadata = { },
+                provider_metadata = arize._generated.api_client.models.ai_integration_provider_metadata.AiIntegration_provider_metadata(),
                 scopings = [
                     arize._generated.api_client.models.ai_integration_scoping.AiIntegrationScoping(
                         organization_id = '',