+++ src/arize/_generated/api_client/test/test_prompt_with_version.py
@@ -97,7 +97,6 @@ class TestPromptWithVersion(unittest.TestCase):
                         anthropic_version = '', 
                         bedrock_options = arize._generated.api_client.models.provider_params_bedrock_options.ProviderParams_bedrock_options(
                             use_converse_endpoint = True, ), 
-                        custom_provider_params = arize._generated.api_client.models.custom_provider_params.custom_provider_params(), 
                         region = '', ), 
                     created_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                     created_by_user_id = '', 
@@ -168,7 +167,6 @@ class TestPromptWithVersion(unittest.TestCase):
                         anthropic_version = '', 
                         bedrock_options = arize._generated.api_client.models.provider_params_bedrock_options.ProviderParams_bedrock_options(
                             use_converse_endpoint = True, ), 
-                        custom_provider_params = arize._generated.api_client.models.custom_provider_params.custom_provider_params(), 
                         region = '', ), 
                     created_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                     created_by_user_id = '',